Generate fake data in integration tests
